Update spec for the libspec branch and add in previously missing elements
This commit has several changes, I moved the template into a separate
file to make it easier to load different versions of the seqspec
template.

I had to update it for the new libspec branch which added new
attributes and renamed some previous attributes.

And I discovered some more molecular elements for the parse
constructs.