Remove .fastq.gz suffix from read_ids
This ended up being needed because I was using a bit of the url to
figure out what the read name was when doing one of my qc checks.

Now that we were capturing many more terms I could look at the
accession instead of href and remove the filename parsing.